+++ diginode-setup.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#           Name:  DigiNode Setup v0.9.10
+#           Name:  DigiNode Setup v0.9.11
 #
 #        Purpose:  Install and manage a DigiByte Node and DigiAsset Node via the linux command line.
 #          
@@ -16,7 +16,7 @@
 #                  Bluesky  - https://bsky.app/profile/diginode.tools
 #                  X        - https://twitter.com/diginodetools
 #
-#    Get Started:  curl http://setup.diginode.tools | bash  
+#    Get Started:  curl -sSL setup.diginode.tools | bash 
 #  
 #                  Alternatively clone the repo to your home folder:
 #
@@ -10676,6 +10676,14 @@ if [ "$DGB_DO_INSTALL" = "YES" ]; then
         str="Renaming 8.22.0-rc3 download folder..."
         printf "%b %s" "${INFO}" "${str}"
         sudo -u $USER_ACCOUNT mv $USER_HOME/digibyte-af6d4e3cdef0 $USER_HOME/digibyte-8.22.0-rc3
+        printf "%b%b %s Done!\\n" "${OVER}" "${TICK}" "${str}"
+    fi
+
+    # WORKAROUND: If this is 8.22.0-rc4, we need to manually rename the extracted directory name, since it is incorrect
+    if [ -d "$USER_HOME/digibyte-527219d69dd9" ] && [ "$DGB_VER_GITHUB" = "8.22.0-rc4" ]; then
+        str="Renaming 8.22.0-rc4 download folder..."
+        printf "%b %s" "${INFO}" "${str}"
+        sudo -u $USER_ACCOUNT mv $USER_HOME/digibyte-527219d69dd9 $USER_HOME/digibyte-8.22.0-rc4
         printf "%b%b %s Done!\\n" "${OVER}" "${TICK}" "${str}"
     fi