+++ diginode.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#           Name:  DigiNode Dashboard v0.9.10
+#           Name:  DigiNode Dashboard v0.9.11
 #
 #        Purpose:  Monitor and manage the status of you DigiByte Node and DigiAsset Node.
 #          
@@ -16,7 +16,7 @@
 #                  Bluesky -  https://bsky.app/profile/digibyte.help
 #                  X -        https://twitter.com/diginodetools
 #
-#    Get Started:  curl http://setup.diginode.tools | bash  
+#    Get Started:  curl -sSL setup.diginode.tools | bash
 #  
 #                  Alternatively clone the repo to your home folder:
 #
@@ -60,8 +60,8 @@
 # Whenever there is a new release, this number gets updated to match the release number on GitHub.
 # The version number should be three numbers seperated by a period
 # Do not change this number or the mechanism for installing updates may no longer work.
-DGNT_VER_LOCAL=0.9.10
-# Last Updated: 2024-02-04
+DGNT_VER_LOCAL=0.9.11
+# Last Updated: 2024-03-29
 
 # This is the command people will enter to run the install script.
 DGNT_SETUP_OFFICIAL_CMD="curl -sSL setup.diginode.tools | bash"